Finish DMA implementation + commands support
The DMA can now be used to load the ".data".

Alternatively, the ".text" and ".data" can be used programmed by first
copying to the p2p region and then issuing a command that moves the data
from the p2p region to the appropriate regions in the device.

There are three other commands as well in this patch:
- Run program: start execution of the eBPF program
- Dump memory: dumps the content of the program and data segments
(currently, this only prints in the screen, it is not sent back to the
guest).

Signed-off-by: Martin Ichilevici de Oliveira <martin.i.oliveira@gmail.com>